backend/project/requirements.txt
asgiref==3.2.10
cffi==1.14.0
cryptography==2.9.2
Django==3.0.8
django-split-settings==1.0.1
djangorestframework==3.11.0
mysqlclient==2.0.1
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
